requirements.txt
# 核心依赖
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.0.1
pymilvus>=2.4.0
milvus-lite>=2.4.4
zhipuai>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
pydantic>=2.0.0